Update objective function for PID fit
Instead of minimizing error starting from a fixed time 't_opt', the objective
function is now defined in terms of the 1) overshoot, 2) rise time, 3) settling
time, and 4) steady state error.

For now, the rise time is hard-coded to zero (despite the `params` file), as it
is the least important of the four.

TO DO: Calculation of the rise time.