Add explicit reset of auxiliary arrays
The auxiliary arrays should be cleared and reset with each call of
`fit_paramsPID()`. I discovered that, previously, this was being done only
serendipitously through `rewindTime()`, as time restarts from zero.

Although this commit presents no change in functionality, it enforces the
intention of the design, which requires the reset of the auxiliary arrays.