Add 'time_lag' parameter for heater
Sometimes it appears that there's a phase offset between the heater and
thermistor, i.e., the thermistor takes a few seconds before responding to the
heater turning on/off. Although this was supposed to be taken care of by the
(thermal) circuit parameters, they don't seem to work as well as the simple lag
term introduced here.